division of logic
