+++ app/quiz/models.py
@@ -2,6 +2,7 @@ from sqlalchemy import Column, String, Integer, ForeignKey, Boolean
 from sqlalchemy.orm import relationship
 
 from db.models import BaseModel
+from quiz.crud import QuestionCrud, QuizCrud
 
 
 class AnswerModel(BaseModel):
@@ -14,7 +15,7 @@ class AnswerModel(BaseModel):
     question = relationship("QuestionModel", back_populates="answers", lazy="subquery")
 
 
-class QuestionModel(BaseModel):
+class QuestionModel(BaseModel, QuestionCrud):
     __tablename__ = "question"
 
     question = Column(String, nullable=False)
@@ -23,32 +24,8 @@ class QuestionModel(BaseModel):
     quiz = relationship("QuizModel", back_populates="questions", lazy="subquery")
     answers = relationship("AnswerModel", cascade="all, delete-orphan", back_populates="question", lazy="subquery")
 
-    async def create(self, db, data):
-        await super().create(db)
 
-        for answer in data:
-            new_answer = AnswerModel(answer=answer.answer, id_question=self.id, is_correct=answer.is_correct)
-
-            await new_answer.create(db)
-
-        await db.refresh(self)
-
-    async def update_with_answers(self, db, data):
-        await self.update(db, {'question': data.question})
-
-        for answer in self.answers:
-            await answer.delete(db)
-
-        for answer in data.answers:
-            new_answer = AnswerModel(answer=answer.answer, id_question=self.id, is_correct=answer.is_correct)
-
-            await new_answer.create(db)
-
-        await db.refresh(self)
-
-
-
-class QuizModel(BaseModel):
+class QuizModel(BaseModel, QuizCrud):
     __tablename__ = "quiz"
 
     name = Column(String, nullable=False)
@@ -58,18 +35,3 @@ class QuizModel(BaseModel):
 
     company = relationship("CompanyModel", back_populates="quizzes", lazy="subquery")
     questions = relationship("QuestionModel", cascade="all, delete-orphan", back_populates="quiz", lazy="subquery")
-
-    async def create(self, db, data):
-        await super().create(db)
-
-        for question in data.questions:
-            new_question = QuestionModel(question=question.question, id_quiz=self.id)
-
-            await new_question.create(db, question.answers)
-
-    async def add_question(self, db, data):
-        new_question = QuestionModel(question=data.question, id_quiz=self.id)
-
-        await new_question.create(db, data.answers)
-
-        await db.refresh(self)
